requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
nltk
beautifulsoup4
tqdm
matplotlib
wordcloud
